Switch to float16 for T4 GPU experimentation

+++ train_gpt2.py
@@ -245,7 +245,7 @@ for step in step_it:
         while xgen.size(1) < max_length:
             # Forward the model to get the logits
             with torch.no_grad():
-                with torch.autocast(device_type=device_type, dtype=torch.bfloat16):
+                with torch.autocast(device_type=device_type, dtype=torch.float16):
                     logits, loss = model(xgen) # (B, T, vocab_size)
                 # Take the logits at the last position
                 logits = logits[:, -1, :] # (B, vocab_size)
@@ -277,7 +277,7 @@ for step in step_it:
         # Added after video, this field is also used by the forward pass.
         if ddp:
             model.require_backward_grad_sync = (micro_step == grad_accum_steps - 1)
-        with torch.autocast(device_type=device_type, dtype=torch.bfloat16):
+        with torch.autocast(device_type=device_type, dtype=torch.float16):
             logits, loss = model(x, y)
         # We have to scale the loss to account for gradient accumulation,
         # because the gradients just add on each successive backward().
